[NEW] frontend-common:  Move non-Mac amicus code
While amicus is a Mac-specific library (which supports a number of macOS
applications whose names all happen to begin with "Am-"), various things
it uses aren't Mac-dependent at all, requiring only a POSIX environment.

+++ graphics/frontend-common/frend/make_cursor.cc
@@ -1,9 +1,9 @@
 /*
-	make_raster.cc
+	make_cursor.cc
 	--------------
 */
 
-#include "amicus/make_raster.hh"
+#include "frend/make_cursor.hh"
 
 // POSIX
 #include <fcntl.h>
@@ -11,17 +11,22 @@
 
 // Standard C
 #include <errno.h>
+#include <stdlib.h>
 #include <string.h>
+
+// shared_cursor
+#include "shared_cursor/cursor_file.hh"
 
 // rasterlib
 #include "raster/load.hh"
-#include "raster/relay_detail.hh"
 #include "raster/size.hh"
 #include "raster/skif.hh"
-#include "raster/sync.hh"
+
+// frend
+#include "frend/cursor.hh"
 
 
-namespace amicus
+namespace frend
 {
 
 static
@@ -46,19 +51,17 @@ int create_raster_file( const char* path, raster::raster_load& result )
 		return errno;
 	}
 	
-	const uint32_t width  = 512;
-	const uint32_t height = 342;
+	const uint32_t width  = 16;
+	const uint32_t height = 16;
 	const uint32_t weight = 1;
 	
-	const uint32_t frame_count = 2;
+	const uint32_t frame_count = 3;
 	
-	const uint32_t stride     = make_stride( width, weight );
+	const uint32_t stride     = 2;
 	const uint32_t image_size = height * stride;
 	const uint32_t raster_size = image_size * frame_count;
 	
 	const uint32_t footer_size_minimum = sizeof (raster_metadata)
-	                                   + sizeof (raster_note)
-	                                   + sizeof (sync_relay)
 	                                   + sizeof (uint32_t);
 	
 	const off_t file_size = good_file_size( raster_size, footer_size_minimum );
@@ -81,7 +84,7 @@ int create_raster_file( const char* path, raster::raster_load& result )
 	
 	const uint32_t footer_size = file_size - raster_size;
 	
-	memset( raster.addr, '\xFF', raster.size );
+	memset( raster.addr, '\x00', raster.size );
 	
 	raster.meta = (raster_metadata*) ((char*) raster.addr + raster_size);
 	
@@ -93,7 +96,6 @@ int create_raster_file( const char* path, raster::raster_load& result )
 	raster_metadata& meta = *raster.meta;
 	
 	raster_desc& desc = meta.desc;
-	raster_note& note = meta.note;
 	
 	memset( &meta, '\0', sizeof meta );
 	
@@ -105,39 +107,39 @@ int create_raster_file( const char* path, raster::raster_load& result )
 	desc.model  = Model_monochrome_paint;
 	desc.extra  = frame_count - 1;
 	
-	raster_note* next_note = &note;
-	
-	next_note->type = Note_sync;
-	next_note->size = sizeof (sync_relay);
-	
-	sync_relay& sync = data< sync_relay >( *next_note );
-	
-	memset( &sync, '\0', sizeof (sync_relay) );
-	
-	next_note = next( next_note );
-	
-	next_note->type = Note_end;
-	
 	result = raster;
 	
 	return 0;
 }
 
-raster_lifetime::raster_lifetime( const char* raster_path )
+cursor_lifetime::cursor_lifetime( const char* cursor_path )
 {
-	its_path = raster_path;
+	its_path = cursor_path;
 	
-	int errnum = create_raster_file( raster_path, its_load );
-	
-	if ( errnum )
+	if ( cursor_path )
 	{
-		throw raster_creation_error();
+		int errnum = create_raster_file( cursor_path, its_load );
+		
+		if ( errnum )
+		{
+			throw cursor_creation_error();
+		}
+		
+		frend::cursor_state = open_cursor_file( cursor_path );
+		
+		if ( frend::cursor_state )
+		{
+			setenv( "AMS_CURSOR", "hardware", 1 );
+		}
 	}
 }
 
-raster_lifetime::~raster_lifetime()
+cursor_lifetime::~cursor_lifetime()
 {
-	unlink( its_path );
+	if ( its_path )
+	{
+		unlink( its_path );
+	}
 }
 
 }
